fix config UI32 get declaration

+++ d912pxy/d912pxy_config.cpp
@@ -136,7 +136,7 @@ UINT64 d912pxy_config::GetValueUI64(d912pxy_config_value val)
 	return _wtoi64(&data[val].value[0]);
 }
 
-UINT64 d912pxy_config::GetValueUI32(d912pxy_config_value val)
+UINT32 d912pxy_config::GetValueUI32(d912pxy_config_value val)
 {
 	return (UINT32)GetValueUI64(val);
 }
